fixed jest errors
